+++ src/main/java/racingcar/controller/RacingCarWebController.java
@@ -4,8 +4,8 @@ import org.springframework.web.bind.annotation.GetMapping;
 import org.springframework.web.bind.annotation.PostMapping;
 import org.springframework.web.bind.annotation.RequestBody;
 import org.springframework.web.bind.annotation.RestController;
-import racingcar.domain.GameInforamtionDto;
-import racingcar.domain.GameResultDto;
+import racingcar.dto.GameInforamtionDto;
+import racingcar.dto.GameResultDto;
 import racingcar.service.RacingCarService;
 
 import java.util.List;
