+++ src/main/java/racingcar/dto/GameInforamtionDto.java
@@ -1,4 +1,4 @@
-package racingcar.domain;
+package racingcar.dto;
 
 public class GameInforamtionDto {
