+++ src/main/java/racingcar/dto/GameResultDto.java
@@ -1,4 +1,4 @@
-package racingcar.domain;
+package racingcar.dto;
 
 import java.util.List;
