+++ src/test/java/racingcar/controller/RacingCarRacingCarConsoleControllerTest.java
@@ -8,7 +8,7 @@ import org.springframework.boot.test.context.SpringBootTest;
 import org.springframework.boot.web.server.LocalServerPort;
 import org.springframework.http.HttpStatus;
 import org.springframework.http.MediaType;
-import racingcar.domain.GameInforamtionDto;
+import racingcar.dto.GameInforamtionDto;
 
 import static org.hamcrest.core.Is.is;
