+++ src/test/java/racingcar/repository/ResultDaoTest.java
@@ -1,4 +1,4 @@
-package racingcar.dao;
+package racingcar.repository;
 
 import org.junit.jupiter.api.BeforeEach;
 import org.junit.jupiter.api.DisplayName;
